move baker code into own file in utils/sv_obj_baker.py
