+++ utils/__init__.py
@@ -112,7 +112,7 @@ def app_handler_ops(append=None, remove=None):
     #     print('error while listing event handlers', err)
     #     listed = ""
 
-    print(f'sv: {operation} app.handlers:')
+    print(f'sv: {operation} app.handlers')
     # print(f'{listed}')
 
 
@@ -120,7 +120,7 @@ utils_modules = [
     # non UI tools
     "cad_module", "cad_module_class", "sv_bmesh_utils", "sv_viewer_utils", "sv_curve_utils",
     "voronoi", "sv_script", "sv_itertools", "script_importhelper", "sv_oldnodes_parser",
-    "csg_core", "csg_geom", "geom", "sv_easing_functions", "sv_text_io_common",
+    "csg_core", "csg_geom", "geom", "sv_easing_functions", "sv_text_io_common", "sv_obj_baker",
     "snlite_utils", "snlite_importhelper", "context_managers", "sv_node_utils", "sv_noise_utils",
     "profile", "logging", "testing", "sv_prefs", "sv_requests", "sv_examples_utils",
     # UI text editor ui
